roaring: cast len to uint64 for 386 build
`len(cfg) > 0xffff_ffff` doesn't compile on 32-bit targets — Go's
untyped int constant 4294967295 overflows int32. Cast to uint64
so the comparison works on every arch (on 32-bit the check is
always false; on 64-bit it stays correct).

+++ pkg/plugin/roaring/roaring.go
@@ -120,7 +120,7 @@ func (c *Cohort) Serialize() ([]byte, error) {
 		return nil, err
 	}
 	out := make([]byte, 4+len(cfg)+len(bm))
-	if len(cfg) > 0xffff_ffff {
+	if uint64(len(cfg)) > 0xffff_ffff {
 		return nil, errors.New("roaring: config too large")
 	}
 	// little-endian length so it matches the Roaring serialization's
